+++ CHANGES.md
@@ -4,6 +4,7 @@
 
 ### Added
 - Partial support for constrained pointer fields [#58](https://github.com/OCamlPro/seacoral/pull/58)
+- Argument `--initialization` to `seacoral check` to assess proper project initialization after configuration checks [#73](https://github.com/OCamlPro/seacoral/pull/73)
 - New subcommand for replaying tests only [#47](https://github.com/OCamlPro/seacoral/pull/47)
 - Association between labels and tests in the corpus [#53](https://github.com/OCamlPro/seacoral/pull/53)
 - A pretty error message when entrypoints or cover targers are missing from the C code [#52](https://github.com/OCamlPro/seacoral/pull/52)